+++ index.html
@@ -57,6 +57,7 @@
     </script>
     <script type="module" src="snow.js"></script>
     <script type="module" src="rain.js"></script>
+    <script type="module" src="script.js"></script>
 </body>
 
 </html>
